# Avoid slices in entries of decomposition tables