@glandium 2022-06-23
Forgot to mention in the commit: relocatable data also means it's not sharable between processes.