@glandium 2022-06-23
With a hacked up cargo-bloat:
on master:
```
$ cargo bloat --release -n 0 --crates
    Finished release [optimized] target(s) in 0.00s
    Analyzing target/release/foors

File  .text     Size Crate
5.1%  96.6% 221.4KiB std
0.0%   0.6%   1.4KiB unicode_normalization
0.0%   0.5%   1.1KiB tinyvec
0.0%   0.1%     339B foors
0.0%   0.1%     229B [Unknown]
5.3% 100.0% 229.1KiB .text section size, the file size is 4.2MiB

Note: numbers above are a result of guesswork. They are not 100% correct and never will be.
$ cargo bloat --release -n 0 --crates --symbols-section .rodata
    Finished release [optimized] target(s) in 0.00s
    Analyzing target/release/foors

File  .text    Size Crate
1.2%  63.7% 51.3KiB unicode_normalization
0.0%   1.0%    851B std
0.0%   0.2%    174B [Unknown]
1.9% 100.0% 80.5KiB .rodata section size, the file size is 4.2MiB

Note: numbers above are a result of guesswork. They are not 100% correct and never will be.
$ cargo bloat --release -n 0 --crates --symbols-section .data.rel.ro
    Finished release [optimized] target(s) in 0.00s
    Analyzing target/release/foors

File  .text     Size Crate
3.1%  93.3% 134.8KiB unicode_normalization
0.0%   0.0%      72B [Unknown]
0.0%   0.0%      16B std
3.3% 100.0% 144.5KiB .data.rel.ro section size, the file size is 4.2MiB

Note: numbers above are a result of guesswork. They are not 100% correct and never will be.
$ cargo bloat --release -n 0 --crates --symbols-section .rela.dyn
    Finished release [optimized] target(s) in 0.00s
    Analyzing target/release/foors

File  .text     Size Crate
3.5% 100.0% 152.7KiB .rela.dyn section size, the file size is 4.2MiB

Note: numbers above are a result of guesswork. They are not 100% correct and never will be.
```
with the patch applied:
```
$ cargo bloat --release -n 0 --crates
    Finished release [optimized] target(s) in 0.00s
    Analyzing target/release/foors

File  .text     Size Crate
5.8%  96.6% 221.4KiB std
0.0%   0.7%   1.6KiB unicode_normalization
0.0%   0.5%   1.1KiB tinyvec
0.0%   0.1%     339B foors
0.0%   0.1%     229B [Unknown]
6.0% 100.0% 229.3KiB .text section size, the file size is 3.7MiB

Note: numbers above are a result of guesswork. They are not 100% correct and never will be.
$ cargo bloat --release -n 0 --crates --symbols-section .rodata
    Finished release [optimized] target(s) in 0.00s
    Analyzing target/release/foors

File  .text     Size Crate
2.8%  83.1% 105.6KiB unicode_normalization
0.0%   0.7%     851B std
0.0%   0.1%     174B [Unknown]
3.3% 100.0% 127.0KiB .rodata section size, the file size is 3.7MiB

Note: numbers above are a result of guesswork. They are not 100% correct and never will be.
$ cargo bloat --release -n 0 --crates --symbols-section .data.rel.ro
    Finished release [optimized] target(s) in 0.00s
    Analyzing target/release/foors

File  .text   Size Crate
0.0%   0.7%    72B [Unknown]
0.0%   0.5%    48B unicode_normalization
0.0%   0.2%    16B std
0.3% 100.0% 9.7KiB .data.rel.ro section size, the file size is 3.7MiB

Note: numbers above are a result of guesswork. They are not 100% correct and never will be.
$ cargo bloat --release -n 0 --crates --symbols-section .rela.dyn
    Finished release [optimized] target(s) in 0.00s
    Analyzing target/release/foors

File  .text    Size Crate
0.5% 100.0% 17.9KiB .rela.dyn section size, the file size is 3.7MiB

Note: numbers above are a result of guesswork. They are not 100% correct and never will be.
```