@Manishearth 2022-06-23
No but I can if you make a PR that bumps the version